Restructure federated modules into dedicated top-level directory for better visibility
