+++ application/account-management/WebApp/rsbuild.config.ts
@@ -21,9 +21,9 @@ export default defineConfig({
     DevelopmentServerPlugin({ port: 9101 }),
     ModuleFederationPlugin({
       exposes: {
-        "./AvatarButton": "./shared/components/AvatarButton.tsx",
-        "./SupportButton": "./shared/components/support/SupportButton.tsx",
-        "./SharedSideMenu": "./shared/components/sideMenu/SharedSideMenu.tsx",
+        "./AvatarButton": "./federated-modules/AvatarButton.tsx",
+        "./SupportButton": "./federated-modules/support/SupportButton.tsx",
+        "./FederatedSideMenu": "./federated-modules/sideMenu/FederatedSideMenu.tsx",
         "./translations/en-US": "./shared/translations/locale/en-US.ts",
         "./translations/da-DK": "./shared/translations/locale/da-DK.ts",
         "./translations/nl-NL": "./shared/translations/locale/nl-NL.ts"
