+++ application/back-office/WebApp/routes/back-office/index.tsx
@@ -2,7 +2,7 @@ import { TopMenu } from "@/shared/components/topMenu";
 import { Trans } from "@lingui/react/macro";
 import { AppLayout } from "@repo/ui/components/AppLayout";
 import { createFileRoute } from "@tanstack/react-router";
-import SharedSideMenu from "account-management/SharedSideMenu";
+import FederatedSideMenu from "account-management/FederatedSideMenu";
 
 export const Route = createFileRoute("/back-office/")({
   component: Home
@@ -11,7 +11,7 @@ export const Route = createFileRoute("/back-office/")({
 export default function Home() {
   return (
     <>
-      <SharedSideMenu currentSystem="back-office" />
+      <FederatedSideMenu currentSystem="back-office" />
       <AppLayout topMenu={<TopMenu />}>
         <h1>
           <Trans>Welcome to the Back Office</Trans>
